editoast: views: move `Ordering` in `operational_studies::hierarchy`
Work schedules endpoint wrongfully uses this enum as well. This should
be addressed at a later time.
