+++ editoast/src/views/operational_studies/hierarchy/project.rs
@@ -18,9 +18,9 @@ use thiserror::Error;
 use utoipa::IntoParams;
 use utoipa::ToSchema;
 
+use super::OperationalStudiesOrderingParam;
 use crate::error::InternalError;
 use crate::error::Result;
-use crate::views::operational_studies::OperationalStudiesOrderingParam;
 use crate::views::pagination::PaginatedList;
 use crate::views::pagination::PaginationQueryParams;
 use crate::views::pagination::PaginationStats;
